[cifuzz][NFC] Rename cifuzz.py to build_fuzzers.py
Since cifuzz.py now only deals with building fuzzers, change the
name to build_fuzzers.py
